Make BOM builder plugin usable with Maven3 as well (#14)
This plugin so far was able to generate BOMs with classifier only, but those are consumable ONLY with Maven4. Goal is to make this plugin useful for Maven3 users as well.

Hence, new "mode" added: when `attach` set to `true`, but no `bomClassifier` is set, then then generated BOM will **replace** current project POM, possibly inheriting some properties of it (but content is the generated BOM). For this to happen, the "bom" subproject must fulfil some criteria: it **must be**:
* packaging=pom
* but must NOT have subprojects

Fixes #13